Bug#35605595 MySQL shell needs to support GTID tags
Starting in 8.3.0, the server supports GTID with tags. Essentially,
after the UUID component of the GTID and before the range, there might
now be a tag to associate with the GTID. This means that a GTID is now
uniquely identified with UUID+tag (empty tags are valid and represent
the previous, non-tagged GTIDs).

The AdminAPI needs to call several server APIS (i.e: gtid_subtract()) in
order to process and compare GTIDs from different servers, like for
example when joining an instances to a Cluster. But since the primary
and the new instance can have different version, this means that the
new instance (running 8.3.0) can have tagged GTIDs, which the primary
(running a version less than 8.3.0) doesn't know how to process. In this
case, the primary throws an error (malformed GTID), which will abort,
for example, a cluster.addInstance() request.

This patch takes this scenario into account by checking the server
version and if the GTIDs of a joining member are tagged. If the server
doesn't support tagged GTIDs and they are present, then the current
command is aborted and the user informed that, in order to use tagged
GTIDs, all members of a Cluster or ReplicaSet must be at version 8.3.0
or newer.

Change-Id: Ida64d28c9e65446e86473963e9f81343ca7819d5