BUG#35806236 AdminAPI should not persist variables that are set to their defaults
Currently, all variables related to GR and/or InnoDB Cluster are
persisted, regardless if their value is the default one or not. This
happens when configureLocalInstance() or configureInstance() is called
(servers with version 5.7) or simply when creating a Cluster (servers
with version 8.0 or above). This is wasteful and persists variables that
might be deprecated or even removed in later versions, unnecessarily
making it more difficult to upgrade a server.

This patch fixes this issue by trying to check, for a specific
variable, if its default value matches the intended one, and not
persisting it if it does. In servers with version 5.7, there isn't a
foolproof method to obtain a variable's default value, so not all
variables are guaranteed to be checked (but most are).

In servers with version 5.7, after calling configureLocalInstance(), or
in servers with version 8.0 (or newer) after calling createCluster(),
the number of variables persisted should be substantially fewer.

Change-Id: Ic90de748aceb95e8ef9035a5df3898675d1b1d85