Fix non-default sys vars
The patch for BUG#34969740 introduced some tests that change the default
values of some sys vars, mainly "auto_increment_increment". However,
these tests use the main test server, which means that further tests
might break because they don't expect non-defaulted values. This is the
case for "util_import_table_norecord.js" where it expects
"auto_increment_increment" to be 1 (instead of 9).

This patch restores the default values of the sys vars used in the
tests.

Change-Id: I2f8ee8c68060bdee43b37541ee86b87b6079cd17