BUG#35795129 BUG#35782142 BUG#29455103 BUG#29949127
Misc cosmetic fixes and code cleanups

BUG#35795129 Incorrect help for Cluster.describe

The helper text for Cluster.describe() mentions that the output includes
a field named 'version' with the version of each member. That's
incorrect and there's no such field in described().

This patch removes the incorrect information from the helper text.

BUG#35782142 Shell user, api, and cmdline docs missing info on resource naming

The API documentation and helper lacks information about naming
limitations for the Cluster/ReplicaSet/ClusterSet name and instance labels.

This patch extends the API documentation and helper with all the
naming limitations.

BUG#29455103 IMPROVE MESSAGES FOR CONFIGUREINSTANCE WHEN INSTANCE IS VALID

dba.configureInstance() prints two similar messages when an instance is
already ready to be used in a Cluster. For example:

The instance 'localhost:3300' is valid for InnoDB cluster usage.
The instance 'localhost:3300' is already ready for InnoDB cluster usage.

The second message is redundant so this patch removes it while improving
the first one.

BUG#29949127 MENTION CURRENT SERVER VERSION WHEN ADD_INSTANCE() FAILS DUE < 8.017

When addInstance() finds that only clone can be used and the target
instance does not support it because of its version, the command prints
an error indicating clone must be used and that it is available since
8.0.17, however, it does not include the target instance version.

To make the message more clear, this patch includes the target instance
version in the error message.

Change-Id: Id0fdac5699e9f0a0697ebe780fd6107d4513826c