BUG#32091724 AdminAPI must not rely on mysql.slave_master_info and mysql.slave_relay_log_info
Although the mysql.slave_master_info and mysql.slave_relay_log_info
tables are currently externally accessible, their usage is supposed to
be internal, with the expectation that the user queries the
performance_schema table to collect any and all necessary data.

This patch replaces all usages of those two tables in the AdminAPI with
equivalent queries to the performance_schema tables, when the server is
at version 8.4 or newest.

Change-Id: I60b0757d16ab13ce98af8eaa809df20d8d7c8c28