Tests clean-up regarding SUPER privilege:
SUPER has been deprecated in 8.0 so it will eventually be removed, all
logic in Shell was already updated to ensure SUPER is not used in 8.0,
however, some tests weren't updated yet.

This patch updates the remaining tests that make use of SUPER to ensure
it's not used when ran against a version >= 8.0.

Change-Id: I7050ae6c07eb41e4cbb03c9b70b5fa8b59afb27b