BUG#35992885 BUG#35997569
BUG#35992885 .rescan() fails to add unmanaged instance if recovery
account is unsupported

During Cluster.rescan(), a check is made to verify if the recovery
account of a given instance has the correct format. If it hasn't, an
error is thrown ("ERROR: Unsupported recovery account 'rpl_user' has
been found for instance ..." as reported in the issue), meaning that the
account isn't updated. This patch makes sure that rescan properly deals
with accounts with unknown formats, by creating a new user (with the
correct format) and assigning it to the recovery channel.

BUG#35997569 Replica misconfiguration after rescan() +
setPrimaryInstance()

Up until now, the commands ReplicaSet.rescan() and dba.createReplicaSet
(with "adoptFromAR" to 1), didn't change the replication account already
associated with the async replication channel. This means that the
account format would be different than the format of an account created
by the AdminAPI. This could lead to problems as described in
BUG#35997569. This patch makes sure that rescan(), if necessary, creates
new users with the correct format and updates the channel to use that
user instead. Also, when createReplicaSet() is adopting from an existing
AR channel, it now guarantees that said channel is using the correct
account format, by creating a new user if necessary.

Change-Id: Ia8601d72b67e6b31fe1c7cff2432beec11d62397