BUG#35828910 "WARNING: Incorrect recovery account" APPEARS FOR INCORRCET FORMAT USERS POST UPGRADE TO 8.0.34
In the rescan() command, while ensuring that the recovery accounts
match, if the current user doesn't have the expected format, it tries to
use the user with the correct format. But currently, if that user
doesn't exists, the command doesn't create it, and it simply moves on,
ignoring the problem. Hence the reported issue, where the status()
complains with the message, but then running rescan() does nothing. The
upgrade scenario is the easiest way to trigger the issue.

This patch fixes the problem by making sure that, during rescan(), if
the intended user doesn't exist, it's created and the account updated.

Change-Id: I7d1a10f6cc7a4ebacd437c16900da47c3ebae087