BUG#35672985 sandbox start/stop scripts use wrong path
Sandbox deployment searches for the mysqld and mysql tooling binaries in
$PATH to use that path in the start/stop scripts.

However, it does not take into consideration if that path is symlinked.
For the basedir, it will always use the full path of the mysql
instalation. For example:

  basedir: /usr/local/mysql-8.1.0
  binary path: /usr/local/mysql -> /usr/local/mysql-8.1.0

If the path to the binaries is a symlink but the basedir path is not,
when the symlink changes because of, for example, a server upgrade, the
scripts will cause the server to use the wrong basedir. For example:

  basedir: /usr/local/mysql-8.1.0
  binary path: /usr/local/mysql -> /usr/local/mysql-8.2.0

This patch fixes that by ensuring that the canonical path to the
binaries is used, eliminating any symlinks encountered in the path.

Change-Id: I3cc68e1c1bffd9447a39d89b8a895db374c9683b