Fix OCI's manifest tests
Previously, it was not possible to create a file representation of a
file which was missing in the manifest. With this change, this is now
possible, error about non-existing file is delayed until the file is
actually read.

Change-Id: I6193bc3bd3e95898f1c97d574b9cf70fb1deb9d4
(cherry picked from commit 9978204ef84b99765b2355a26e37973ea3476a6b)