BUG#36041691 loadDump ignores leading '0' in S3 bucket prefix
This problem is caused by the CLI mapper, where values are parsed
and the data type is interpreted after the parsing. Numeric values
with leading 0's would be interpreted as numbers, but would loose
the leading 0's.

This patch fixes the problem by adding ensuring numeric values don't
lose data after the parsing, in such case, they are interpreted as
strings.

Change-Id: I648e41e006547abebec5cf5dfb4efeffa89c2323