BUG#30896233 ADDING A (MISSING) INSTANCE RESULTS IN MISLEADING ERROR
Attempting to add an instance to a Cluster, which is already part of the
cluster but is currently (MISSING), is allowed and results in a confusing and
misleading error indicating the instance has a server_uuid already in
use by... the same instance.

This patch fixes that by extending the validations in addInstance() to
consider instances with status (MISSING) to issue a proper error to the
user also indicating that Cluster.rejoinInstance() must be used instead.

Change-Id: Ia3f34bbac65b0486ba02c6181be02cd4fa676070