Tweak exitStateAction default value and remove support for group_replication_recovery_complete_at
With WL15713, in version 8.4.0, the server will change the default value
of 'group_replication_exit_state_action', from "READ_ONLY" to
"OFFLINE_MODE". This patch adjusts the help messages associated to the
corresponding option, "exitStateAction", to reflect this change.

Also, with WL15461, the server removed the deprecated sys var
'group_replication_recovery_complete_at' in version 8.4. This patch also
makes sure that the Shell doesn't test for the existence of said sys var
with 8.4 or newer versions of the server.

Change-Id: Ie4a09f50e9239cdd2f4d58ebcb16a38b70da4ff7