WL#15870 AdminAPI: Removals
Part 2: Removing 5.7 support

Raised the minimum required server version from 5.7 to 8.0 for the
AdminAPI. All commands that previously showed a deprecated warning
message regarding the server version being 5.7 now trigger an exception
informing that the version is no longer supported.

For dba.getCluster(), if the primary of the Cluster is running version
5.7, it now throws an exception informing that the version is not
supported, hence the Cluster can't be retrieved.

Removed and cleaned-up code related to supporting 5.7. Also, when
running the test suite, if the test begins with "js_adminapi" or
"py_adminapi" and doesn't end with "_norecord", the test is now skipped
if the configured target server is 5.7.

Change-Id: I190d9f0b0971fcaee2632de158ef0c56e303937d