BUG#35913824 Review replica-set wide locks
With WL#14870, support for ReplicaSet wide locks (identical to the
already existing ones for Cluster and ClusterSet) was added.

With this support, ReplicaSet commands no longer need to acquire
multiple per-instance locks to simulate ReplicaSet wide locks, which is
far more optimal since it can reduce multiple lock requests to a single
one.

This patch updated the following commands:
 - dba.createReplicaSet: xlock on target instance
 - replicaset.addInstance: xlock on the replicaset and a xlock on the
target instance
 - replicaset.dissolve: xlock on replicaset
 - replicaset.rejoinInstance: slock on replicaset and a xlock on the
target instance
 - replicaset.removeInstance: xlock on the replicaset and a xlock on the
target instance
 - replicaset.rescan: xlock on the replicaset
 - replicaset.setInstanceOption: xlock on the target instance
 - replicaset.setOption: xlock on the replicaset
 - replicaset.setPrimaryInstance: xlock on the replicaset
 - replicaset.forcePrimaryInstance: xlock on the replicaset
 - replicaset.setupAdminAccount: slock on replicaset
 - replicaset.setupRouterAccount: slock on replicaset

This is the same behavior as the equivalent commands in Cluster.

Change-Id: I732cad36e862cae43a386a60b402176edf2314df