WL#16123 Certificate-based authentication support on InnoDB Cluster Read Replicas
InnoDB Cluster Read Replicas were introduced in 8.1.0 (WL#15056),
however, support for certificate-based / passwordless authentication was
not included. This limits the configuration of a Cluster where Read
Replicas can be used, mainly: the Cluster's sslMode is not VERIFY_CA or
VERIFY_IDENTIFY and its memberAuthType must be PASSWORD.

With this patch, full support for the Cluster options "memberSslMode"
and "memberAuthType" are implemented for Read Replicas. This includes a
new option, "certSubject", in cluster.addReplicaInstance().

Also included in this patch are a few tweaks regarding the value of the
"memberAuthType" option in the help of dba.createCluster(),
dba.createReplicaSet(), cluster.addInstance(),
replicaSet.addInstance() and clusterSet.createReplicaCluster() commands.

Change-Id: I20c9b6407a291f5c3ee09d0d4c9f5b92ee8ce847