BUG#36187059 Review usage of group_replication_allow_local_lower_version_join
WL#16018 deprecates "group_replication_allow_local_lower_version_join"
on version 8.4.0. This is because the server now allows lower versions
within the same LTS to join a LTS version, making the option redundant.

This patch removes the usage of the option from AdminAPI if the cluster
minimum version is 8.4.0 or newer. Even if the option is enabled on
servers with version 8.4.0 (where it's deprecated), AdminAPI ignores
it and checks for version compatibility using the established version
rules (as described above).

Change-Id: I2d1c53d44e79b3a7037f2544b929b1deeea4a32c