WL#15432 BUG#36297348 BULK LOAD
WL#15432 BULK LOAD support - util.loadDump

The loader will use the BULK LOAD statements to load the data when all
of the following are true:

 - target instance supports the BULK LOAD statement and version is
   >= 8.4.0
 - target instance has the binlog disabled or the `skipBinlog` option
   is set to true
 - dump is loaded from:
   - the OCI using a prefix PAR and user loading the dump has the
     LOAD_FROM_URL privilege
   - the AWS and user loading the dump has the LOAD_FROM_S3 privilege

In order for a table to be loaded by the BULK LOAD statement, all of
the following must be true:

 - data dump for this table is complete
 - data files are not compressed or are compressed using the `zstd`
   compression
 - table fulfills all requirements to be loaded by the BULK LOAD

If a partitioned table fulfills all of the above conditions, each
partition is loaded by the BULK LOAD.

A new option has been added to the loader:

disableBulkLoad - bool (default: false) - Do not use BULK LOAD
feature to load the data.

BUG#36297348 s3EndpointOverride <override_url> not used for LOAD DATA INFILE S3

When s3EndpointOverride option is used, the BULK LOAD should be
disabled, as it's not possible to convey the overridden endpoint to the
LOAD DATA FROM S3 statement.

Change-Id: If67f6c783372875a71e9836388673354fe091591