Upgrade Checker: condition class hierarchy
So far a unique way was used to determine whether a check should
be enabled or not on the upgrade check process, however, the
need to support different types of conditions arises and so this
class hierarchy will allow introducing new conditions in a
more organized way.

This patch introduces a base Condition interface as well as a
Version_condition class that implements the only condition used
so far.

Change-Id: If9115cd3ac791b489d466187c8d39cac5d19d5e0