BUG#36029413 replicationSslMode option ignored in createReplicaSet(.., {adoptFromAR:1})
WL13688 specifically says that the options "memberAuthType",
"certIssuer", "certSubject" and "replicationSslMode" cannot be used when
adopting either a Cluster or a ReplicaSet (except for "memberAuthType"
with the default "PASSWORD" value).

But the only option actually verified is "memberAuthType". All other
options, when adopting, are ignored. This is incorrect and an error
should be thrown if the user is adopting an existing Cluster or
ReplicaSet and tries to use said options.

This patch adds checks to dba.createCluter() and dba.createReplicaSet(),
if the aforementioned options are use when also adopting an existing
topology, throwing an error if they are.

Change-Id: I9185089c0e43a157c5fb17fdffffa990c65025eb