BUG#36221818 Copy operation hangs
In certain situations, copy operation would hang when target was
constantly scanning for metadata, while source was already waiting to
write the data files. This could happen if chunking tasks were quick to
be executed and produced data tasks, while metadata tasks are still
being queued.

As a fix, metadata tasks are queued first, while chunking tasks are
gathered in a container, and pushed into queue only after the metadata
task are ready.

Change-Id: I2e02f065b273036b4ba47af59a6dce834b0198ea