WL#15974 Upgrade Checker filtering
This worklog introduces a system for including and filtering out
specific checks in Upgrade Checker to better control the whole process.

All checks now print out their identifier that can be used in filtering.

Added "include" option - it will limit all checks to the specified in this identifier list.
Added "exclude" option - it will remove specified checks from performed list.

Both options can be used simultaneously with the condition
of not repeating identifiers in both lists.

Added "list" option - if true, it will print out all checks that would be included
in the process, with its identifier, a short description, conditions that must be
met to be included and then print out all other checks that ware rejected.

"list" option can be used without a session - it will include all possible checks
without rejections.  It also can be combined with "include" and
"exclude" options, to control the resulting list.

Change-Id: I40bc972a33450e2eb6913b2c4999198f08c24e4a