Removes titles, messages and doclinks from the C++ code.
Huge cleanup of the upgrade_checker.msg file:
- Ensures existence/consistency of all messages.
- Adds missing messages for some checks.
- Uncomments most of the messages.

Removes the get_*_internal() functions for titles, messages and doclinks.
Uses constants for the check IDs.

Removes the --generate-uc-translation option available in the debug shell (not used).

Adds logic to enable resolution of tokens within the texts registered in the messages file.

Change-Id: I305b00fe162b41e4dd49c8bde210ab4aa496a2e2