BUG#36057800 AdminAPI uses/sets removed option binlog_transaction_dependency_tracking in 8.4
The AdminAPI verifies and sets a value for
'binlog_transaction_dependency_tracking' as part of the parallel
appliers handling introduced in 8.0.23 (WL#13837).

However, in 8.4.0, Group Replication removes this deprecated sysvar
(WL#13965).
The sysvar was deprecated in 8.0.35/8.2.0.

For that reason, the AdminAPI must not verify and/or set a value for
this sysvar when the target MySQL instance is running a version >=
8.4.0.

Also, Sandboxes are deployed with
binlog_transaction_dependency_tracking pre-configured too.

This patch fixes that by ensuring the commands no longer check and set
binlog_transaction_dependency_tracking when the target version is >=
8.4.0.

Change-Id: I8d9ac3f4688d75da1b10ad5f6a73a5cdb3cab951