Retool API options support
Added two new helper classes to better manage and organize command
options.

The first, "Option_data" allows one to instanciate options with more
data than just the name, like for example, type and scope. With it, we
can more easily organize the options themselves and do not require
multiple inheritance to organize them. We can still do that, but now we
have another option.

The second helper class "Option_pack_builder", is just a wrapper around
the existing "Option_pack_def" to allow the usage of the new
"Option_data". It also decouples the setters / filters / converters from
the object instance (where a group of options are stored), allowing the
same filter or checked to be used by several options without it being
coupled to the object itself. This means that we don't need to use
inheritance to propagate the filter.

With these new tools, re-organized all "dba" and "cluster" level
commands.

Change-Id: I3cd493b184bead57ed4e1e742c4d9d0ed160d6f7