BUG#36229123 RemoveInstance() breaks entire ClusterSet - unrecoverable state
In a scenario where a ClusterSet has multiple clusters, it's possible,
by mistake, to try and remove from one cluster a member belonging to
another.

Altought the AdminAPI performs several checks regarding the instance
being removed, it doesn't check if it belongs to the actual cluster from
which it is being removed, only that it is part of the ClusterSet. This
means, for example, that the user can, by mistake, using a secondary
cluster, remove the primary member of the primary cluster. This leaves
the ClusterSet in an unrecoverable state.

This patch adds this specific check, throwing an error if the member
being removed doesn't belong to the cluster (of the ClusterSet) from
were the command was invoked.

Change-Id: I8e11e946d63e8a1a5ddbabdcde2565f36b434a5b