WL#15954 BUG#36275884 BUG#36271221
AdminAPI: List Router options

Typically, the Router is deployed on the same host as the application
that uses it. This means that users often deploy many instances of the
Router in an InnoDB Cluster/ReplicaSet/ClusterSet environment.

Those deployments are usually automated with tools such as Ansible or
Chef, or even made into microservices using containers. Such
deployments should require no user interaction.

For those reasons, it should be possible to manage Router instances
remotely, and at the moment that's only partially possible for routing
options (WL#14538, WL#15601, WL#15842)

In 8.4.0 (WL#15649), Router exposes all of its configuration in the
Metadata, enabling Shell to consume it and expose it to the users.

This patch adds support for visualizing all Router configurations with
the AdminAPI.

A new command was added, named .routerOptions() and available in all
topology objects (Cluster, ClusterSet, ReplicaSet) that lists the
Router configuration options of the topology (global) and the Router
instances.

By default, only the options that can be changed from Shell
(dynamic options) are displayed. Router instances with different
configurations than the global ones will include the differences under
their dedicated description.

The command supporst the options 'extended' and 'router':

  - extended: Verbosity level of the command output.
  - router: Identifier of the Router instance to be displayed.

The 'extended' options supports the following values:

  - 0: Includes only options that can be changed from Shell (default);
  - 1: Includes all Cluster global options and, per Router, only the
       options that have a different value than the corresponding
       global one.
  - 2: Includes all Cluster and Router options.
  - Boolean: equivalent to assign either 0 (false) or 1 (true).

Change-Id: Ib755b7499da16be44d07d407480a09ec991e88a4