BUG#36225607 dba.rebootClusterFromCompleteOutage() looks for GR on read replica
If a Cluster is rebooted from a session to a read-replica instance or
the "primary" option is used and set to a read-replica, the command
either switches to a valid seed or fails completely. This is correct,
however, in both scenarios, the output shown isn't the most accurate one
(read-replicas aren't even mentioned).

This patch tweaks the checks made in
dba.rebootClusterFromCompleteOutage() to properly deal with
read-replicas and output more accurate info.

Change-Id: Id32c869aa198b82d8990074070c3af279a3a61e3