Rework RPATH handling in case of bundled binaries
Previously, RPATH was written to the original binaries, which were both
used by the installer and copied into build directory. With this change,
binaries are first copied, then RPATH is written to the copy (if
necessary), and copies are also used by the installer. This allows to
keep original files intact.

Change-Id: I5475deafbd111792e93217db50aba318c1566e06