BUG#36276246 Update cluster.removeRouterMetadata() to cleanup cluster(set)s.router_options
In 8.4.0, Router and Shell introduced a mechanism for the Router to
report its configuration and defaults in the Metadata and Shell to
consume, format and display it to the users with .routerOptions().

However, the .removeRouterMetadata() command does not take in
consideration a possible need to clean-up that data. Specifically, for
versions of Metadata >= 2.0 (that will include the new information), if
the Router being removed is the last one of the topology, or the highest
version of the Default configurations version is higher than the Router
with highest version bootstrapped in that topology, the Defaults data
can be removed.

This patch ensures that logic is implemented to perform the clean-up
missing in .removeRouterMetadata().

This patch also fixes a bug in .routerOptions(), when using the
extended:1 option, that "common" options to not be filtered out from
the Default options list.

Change-Id: Ibb3847eaab20cc6ba915c6d74a1ec736cbf55920