BUG#36229274 Read Replica replication source set to other cluster
When adding a Read-Replica to a cluster that belongs to a ClusterSet,
it's possible to specify the replication sources explicitly using the
"replicationSources" parameter. The command checks if the sources
belong to any of the clusters of the ClusterSet, but not if they belong
to the Cluster from which the command is being executed. This means that
the source of a Read-Replica in a replica Cluster, could end up being
an instance of the primary or any other replica Cluster.

This patch adds this check, by not allowing sources that do not belong
to the Cluster where the command was invoked, from being used.

Also fixed a similar issue when calling .addReplicaInstance(),
.rejoinInstance() and .createReplicaCluster(), and the "cloneDonor"
parameter specified an instance from a cluster other than
the target one or the primary one.

Also when calling .setInstanceOption() with "replicationSources", we now
check if the source isn't a read-replica, isn't itself, has a
valid state, is reachable and belongs to the correct cluster.

Change-Id: Ib50721ea6a877fb23dfdd52e4f188805f0d757b8