BUG#36159820 Automatically exclude oracle-cloud-agent user when dumping from/loading into OCI
The following users have been added to the list of user automatically
excluded when loading a dump into an MHS instance:
 - ocidbm
 - oracle-cloud-agent
 - rrhhuser

The users which are automatically excluded when loading into an MHS
instance are now also automatically excluded when dumping with the
`ocimds` option set to true.

A warning is now printed when dumping users, and an account is granted a
role which is not included in the dump.

Change-Id: I513cfae6e4790becc949a16cb42528fa5688bcd9
(cherry picked from commit e3e858fa99e12b7e19c4093791e782dabbedc168)