Post-fix for WL#15954 - AdminAPI: List Router options:
The logic to compute the differences of each Router configuration with
the Defaults checks whether a category exists in both to proceed with
the diff, and if the category does not exist in the Defaults it adds it
to the resulting diff schema.

However, this logic was not verifying sub-options of each category, i.e.
if it's just a sub-option that doesn't exists in a category it'd be
ignored and not shown in the diff schema.

This patch fixes that to ensure any option or sub-option that doesn't
exist in the Defaults is included in the diff schema so it's shown in
each Router configuration.

Change-Id: I663ab59e0a685ae8ae173504ae3384ae0cbffd82