Bug#35751281 mysqlsh displays incorrect connection details
When socket path is specified (in my.cnf or cmdline) but the connection
ends up using TCP because a non-localhost address is set, the connecting
message still showed a socket connection to localhost.

The connection option handling from command line and my.cnf was cleaned
up so that:
- the transport to be used for the connection is determined by the
right-most connection option in the command line (my.cnf options are
at the left-most). If localhost or a socket path is the last, then
socket is used. If port or !localhost is the last, then TCP is used.

- port and socket options are now always preserved, for cases like
	mysqlsh -S/tmp/socket -h127.0.0.1 -hlocalhost
previously, the overriden socket path (and port) would've been lost.

shcore::get_connection_options() becomes redundant because
set_default_data() is set in connect(), where we know all intended
settings were already changed.

Change-Id: Ie1a80769d26bed593592d492dcaaa8de1fbbd26b