BUG#36119568 SHELL LOAD: (Windows) ERROR 1231 (42000): Variable 'wait_timeout' can't be set to the value of '31536000'
Load failed on Windows if the global sql_mode was set to
STRICT_ALL_TABLES.

When loader creates sessions which are used to execute the operation,
the 'wait_timeout' variable is set to 31536000, which is the maximum
allowed value on non-Windows platforms. On Windows, maximum value is
2147483, and when STRICT_ALL_TABLES sql_mode is set, this query fails.
If this mode is not set, 'wait_timeout' variable is set to maximum
possible value, and a warning is issued.

As a fix, the initialization code for each session has been adjusted,
the session's sql_mode is set before executing any other query.

Change-Id: Ie4f6cd23d1415e27c48b0b9d528747d5a4c8e063
(cherry picked from commit 54cedca838450d208cb8d048c607b9b5674deafc)