BUG#36342905 routerOptions() gives no config, routingOptions() does
ClusterSet.routerOptions() may fail to report any configuration
settings when the Routers connected to the ClusterSet have not been
bootstrapped against the ClusterSet but only against the cluster that
was used to build the ClusterSet. In the same situation,
ClusterSet.routingOptions() does display configuration settings.

This issue happens because .routerOptions() expects the changes schema
to be present in the Metadata. The schema is queried to perform
filtering when the 'extended' option is 0 (default) or 1.

This patch fixes that by ensuring the code accepts the scenario on which
there is no changes schema yet because a re-bootstrap is needed, and
returns the default options unchanged. This allows the user to consult
the topology's default options nevertheless.

Change-Id: Ie98a6cfac8e0ec7c1c93b885f8ea84e9dab62957