BUG#36343374 routerOptions(): normalized value (hash) vs. user input (name/option)
Cluster(set).routerOptions() shows a normalized or resolved value for
configuration settings. It does not, however, translate the value for
"target_cluster" (of "routing_options") into a human-readable value.
The value gotten is the Group Replication group's name UUID, which has
no value to the common user.

The deprecated command Cluster(set).routingOptions(), and also
.listRouters(), do the translation of that UUID into the Cluster's name,
which makes it quite convenient for the user. Also, note that
.setRoutingOption() accepts Cluster names for the option
"target_cluster" which is then automatically converted into the
corresponding UUID to be stored in the Metadata.

This patch adds the missing translation.

Change-Id: I43c44b9a19095b26b2e9b056f0ad52df619356e9