BUG#36054489 Relax version compatibility rules for innodb clone usage in AdminAPI
With WL#15989, the compatibility checks of the versions of a donor and
recipient instances during cloning were relaxed. Up until this WL, both
versions had to match (major + minor + path), but now they only require
major + minor (ignoring the patch number in most situation).

The new rules are:
 - only versions >= 8.0.17 can perform cloning
 - if both versions are >= 8.0.37, then only the major and minor
versions need to match (e.g.: 8.1.4 and 8.1.10 is allowed)
 - every other case (>= 8.0.17 and < 8.0.37) require that the major,
minor and patch numbers match.

These apply when the recipient has version 8.4 (or newer) or 8.0.37,
otherwise the previous rule (major, minor and patch numbers must match),
remains the same.

This patch updates the checks made in the AdminAPI to match these new
rules.

Change-Id: If242d0957a9ff8c5032af8bd72e67fca15db0ce5