BUG#36201015 Unable to adopt existing ReplicaSet topology if report_host/report_port not set
The output of the query "show replicas" contains two columns, "host" and
"port", that are used by several commands to then connect to each
replica. But the content of the two columns are the values of the
options " --report-host" and " --report-port" respectively, which means
that, in scenarios where the options weren't configured, the query
returns "" for the host and might return 0 for the port. This means that
connections then fail, or worse, connect to the wrong instance (because
an empty host defaults to the localhost).

This patch fixes this issue by dealing with this scenario: when adopting
an existing ReplicaSet (dba.createReplicaSet), instances without host
and/or port are ignored and the user is informed.

Change-Id: I22512713aa90d355f84dd066cc3b8a4141dce13b