BUG#36298612 BUG#36298109 BUG#36314047 BUG#36236771 Upgrade Checker Bugfixes
- BUG#36298612 Unsetting variable to become valid - enum and "empty string"

The sysvarAllowedValues did not consider the following variables also
accepted an empty string as a valid value, producing false negative errors.

- ssl_cipher
- admin_ssl_cipher
- tls_ciphersuites
- admin_tls_ciphersuites

This patch fixes the problem, by treating en empty value as valid.

- BUG#36298109 Typo: use the targetVersion option... (3x ".")

This patch removes the extra dots.

- BUG#36314047 Update checker sysVarsNewDefaults requires config on 5.7

This patch updates the UC help to indicate the configuration path
is required by the system variable related checks if running with a
server older than 8.0

- BUG#36236771 help does not add anything over shell.connect() help?

This patch replaces the help cbout connection data with an indication
to get additional details using \? connect

Change-Id: I4ab5062cb3a386d000199c509f78eaf1f567414e