BUG#36267549 Cluster.setRoutingOption() accepts router address, no router instance def needed
Most Router management AdminAPI commands accept an optional parameter to
specify a target Router, either to configure it or filter it out from
the output. That Router is specified as a "Router identifier" which has
the following format: <address>::<router_name>.

The commands verifies if the target Router exists in the topology
according to its unique identifier. However, .setRoutingOption()
accepts the address as the Router Name. If a router was configured
without a name, it will be registered in the Metadata as '<address>::'
and the command would accept using <address> as the identified of the
Router. For example:
  c.setRoutingOption("localhost", "read_only_targets", "all");

This is wrong and incoherent with the definition of Router identifier
and the valitions performed by the other commands.

This patch fixes that to ensure the correct validation is done.

Change-Id: I9c9fb90149ea57e10b8b0aaaad0ce3486a0bbdf4