BUG#36338711 Unexpected version found for option support check
The AdminAPI, for most of its operations of topology management,
validates whether an option being used/set is supported by the target
instance. This is externalized, in the most simple scenario, such as
adding an instance to a Cluster.

When that target instance is running an unsupported version, the
AdminAPI throws an error saying "Unexpected version found for option
support check". The error is not clear nor helpful to the user.

This patch improves error handling by providing meaningful messages
indicating that the target instance's version is unsupported because
this version of MySQL Shell supports up to version <max supported>,
or from <min supported> and above.

Change-Id: Ie5431a236e0af569ce7b8b0d287fb26c2ccbd69c