BUG#36359408 Upgrade Checker should warn when targetVersion is higher than the recommended
This commit adds a new general warning in upgrade checker. It is not recommended
(and not supported) to upgrade MySQL Server to very high version in one step.
Upgrade checker now detects the recommended next LTS version to execute an upgrade
and shows a WARNING if the specified target version is bigger than the recommended
ont.

Change-Id: I5cc7f2a3252ce501691f1eff16f75002d9f35372