BUG#36348724 Upgrade checker test is failing with syntax error
The problem was caused by a lost validation, that only checks that
are enabled should be executed, i.e. the failed tests were being
executed even they were marked as disabled, which lead to the syntax
error.

Additionally:

- Added internal condition for listing purposes on
the feature like tests, which don't have a condition associated
when the registration occurs, this is because the condition
actually depends on the registered features.

- Excluded the mysqlSchema check when executing the UC in the context
of a dump operation, the check is pointless as the mysql schema is
never included in the dump.

Change-Id: Id066c57de3b48d6a61bd9ca59f99cf67796514ed