BUG#36381849 MySQL Shell - util.loadDump - Error while executing postamble SQL
In certain situations, the load could fail at the end, when executing
the postamble SQL file. The session used by the main thread lost
connection with the server, after being idle for most of the load
operation. This is unexpected, as loader sets maximum possible wait
timeouts.

As a workaround, when a copy/load query fails with a connection error,
session is reestablished and query is retried.

Statements which are not idempotent are not retried.
Statements which load the data are not retried.

Change-Id: Ie71c73ed8e13563467fff1660524a899373eb2f6