BUG#36348763 Fix mysqlsh default mode so it aligns with the mysql client
MySQL Shell's default mode is JavaScript (JS), which does not align
with the MySQL Client.

This patch changes Shell's default mode to SQL.

NOTE: Up to now, executing JS code from the cmdline didn't require the
usage of '--js' since JS mode was the default. For example:

  $ mysqlsh user@host:3306 -e "println(session)"

With this change, that becomes a requirement:

  $ mysqlsh user@host:3306 --js -e "println(session)"

Change-Id: I7a7f232757215df468df2c82970a75f592eb7a5f