BUG#35387205 Cannot rejoin instance in ERROR state which failed on distributed recovery
Rejoining an instance in ERROR state, with .rejoinInstance(), fails
with errors indicating that Group Replication settings cannot be
changed when Group Replication is running. This is correct from Group
Replication's point of view but the information given to the user is
vague, incomplete and does not include a solution.

Considering that a member can go into ERROR state due to one of the
following:

  - Error on the group_replication_applier channel
  - Error on the group_replication_recovery channel
  - Internal MySQL/Group_Replication errors (memory, config access,
    etc.)

...being the most common scenario a member that failed to complete the
recovery or failed to apply some transaction.

...and that the only way to overcome that is to STOP Group Replication
or to restart MySQL.

We can safely assume that .rejoinInstance() should first stop
Group Replication if the target Cluster member is in ERROR state.

This patch applies that logic (stop GR) if the instance is in ERROR
state so that the command can proceed with the rejoin process.

Change-Id: Id789a8a895b08b3049187407b5fb848e07e68ca2